+++ src/routes/(app)/eventos/[slug]/+page.svelte
@@ -191,6 +191,7 @@
                   <AttendanceStat
                     assistance={event.assistance}
                     event={event.title}
+                    venue={event.venue?.venueName}
                   />
                 {/if}
               {/if}