+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Nav button names
 - Overflow of Team Cards on Cartelera component
 - Fix bottom margin on WhoWeAre component
+- Footer year and attendace stat component UI
 
 ## [0.0.87] - 2025-01-14